Use permission modding in both export tasks.
